+++ cpp/include/cudf/detail/utilities/linked_column.hpp
